feat: introduce ModelToolbar and PolicyToolbar components
